fix(cache): address review feedback on entity change log
- Move OperateType and EntityChangeRecord out of po.auth into po.cache;
  they are general cache-invalidation infra, not authn/authz POs.
- Rename selectChanges parameter from createdAtAfter to createdAtFrom
  to match the inclusive (>=) semantics already in the SQL.
- Use sharedSession for @BeforeEach truncate in TestEntityChangeLogMapper
  and clear the local cache so MyBatis SESSION-scoped cache cannot
  serve stale results across tests.
