--- core/src/main/java/org/apache/gravitino/storage/relational/po/auth/EntityChangeRecord.java
@@ -1,95 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.gravitino.storage.relational.po.auth;
-
-/** Entity change poller result -- one row per entity_change_log entry. */
-public class EntityChangeRecord {
-  private long id;
-  private String metalakeName;
-  private String entityType;
-  private String fullName;
-  private OperateType operateType;
-  private long createdAt;
-
-  /** Required by MyBatis for result mapping. */
-  public EntityChangeRecord() {}
-
-  public EntityChangeRecord(
-      long id,
-      String metalakeName,
-      String entityType,
-      String fullName,
-      OperateType operateType,
-      long createdAt) {
-    this.id = id;
-    this.metalakeName = metalakeName;
-    this.entityType = entityType;
-    this.fullName = fullName;
-    this.operateType = operateType;
-    this.createdAt = createdAt;
-  }
-
-  public long getId() {
-    return id;
-  }
-
-  public void setId(long id) {
-    this.id = id;
-  }
-
-  public String getMetalakeName() {
-    return metalakeName;
-  }
-
-  public void setMetalakeName(String metalakeName) {
-    this.metalakeName = metalakeName;
-  }
-
-  public String getEntityType() {
-    return entityType;
-  }
-
-  public void setEntityType(String entityType) {
-    this.entityType = entityType;
-  }
-
-  public String getFullName() {
-    return fullName;
-  }
-
-  public void setFullName(String fullName) {
-    this.fullName = fullName;
-  }
-
-  public OperateType getOperateType() {
-    return operateType;
-  }
-
-  public void setOperateType(OperateType operateType) {
-    this.operateType = operateType;
-  }
-
-  public long getCreatedAt() {
-    return createdAt;
-  }
-
-  public void setCreatedAt(long createdAt) {
-    this.createdAt = createdAt;
-  }
-}
